# Docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AUSTA Cockpit API
  description: |
    Medical Audit Platform API with AI-powered decision support.
    
    This API provides endpoints for managing medical authorization cases, 
    AI-assisted analysis, audit decisions, and comprehensive analytics.
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Obtain a token through the 
    `/auth/login` endpoint and include it in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API calls are rate limited to ensure fair usage:
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour
    - Premium: 10000 requests/hour
    
    ## Environments
    
    - Development: `http://localhost:3000/api/v1`
    - Staging: `https://staging-api.austa.com.br/v1`
    - Production: `https://api.austa.com.br/v1`
    
  version: 1.0.0
  contact:
    name: AUSTA Support
    email: suporte@austa.com.br
    url: https://docs.austa.com.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://staging-api.austa.com.br/v1
    description: Staging server
  - url: https://api.austa.com.br/v1
    description: Production server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            requestId:
              type: string
              format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...